refactor filenames (B11)
