+++ src/components/additional-film-blocks.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from "./abstractComponent";
+import {AbstractComponent} from "./abstract-component";
 
 export class AdditionalFilmList extends AbstractComponent {
   constructor(blockName) {
