+++ src/components/default-film-list.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from "./abstractComponent";
+import {AbstractComponent} from "./abstract-component";
 
 export class DefaultFilmList extends AbstractComponent {
   constructor() {
