+++ src/components/empty-search.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from "./abstractComponent";
+import {AbstractComponent} from "./abstract-component";
 export class EmptySearch extends AbstractComponent {
   constructor() {
     super();
