+++ src/components/sort.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from "./abstractComponent";
+import {AbstractComponent} from "./abstract-component";
 
 export class Sort extends AbstractComponent {
   constructor() {
