+++ src/components/statistics-rank.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from "./abstractComponent";
+import {AbstractComponent} from "./abstract-component";
 
 export class StatsRank extends AbstractComponent {
   constructor(rank) {
@@ -8,8 +8,8 @@ export class StatsRank extends AbstractComponent {
 
   getTemplate() {
     return `<p class="statistic__rank">
-      Your rank 
-      <img class="statistic__img" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35"> 
+      Your rank
+      <img class="statistic__img" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
       <span class="statistic__rank-label">${this._rank}</span>
     </p>`;
   }
