+++ src/components/statistics-section.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from "./abstractComponent";
+import {AbstractComponent} from "./abstract-component";
 
 export class StatisticsSection extends AbstractComponent {
   constructor() {
